# services/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Route inventory with explicit operations schemas for the admin client. Unspecified payloads remain unknown.",
    "title": "TripRaft API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/group-planner/groups/{group_id}/calendar.ics": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_export_export_ical_get_api_v1_group_planner_groups_group_id_calendar_ics",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/group-planner/groups/{group_id}/clone": {
      "post": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_export_clone_group_post_api_v1_group_planner_groups_group_id_clone",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/group-planner/groups/{group_id}/export/pdf": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_export_export_pdf_get_api_v1_group_planner_groups_group_id_export_pdf",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
  }
}